requirements.txt
asgiref==3.7.2
Django==5.0.1
djangorestframework==3.14.0
Faker==22.6.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
